diffusion-ui/requirements.txt
gradio>=5.22.0
Pillow>=10.4.0
requests>=2.32.0
